css/css-color/parsing/color-computed-named-color.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Color Level 4: Computation of colors using named color notation</title>
<link rel="help" href="https://drafts.csswg.org/css-color-4/#named-colors">
<link rel="help" href="https://drafts.csswg.org/css-color-4/#resolving-sRGB-values">
<link rel="help" href="https://drafts.csswg.org/css-color-4/#serializing-sRGB-values">
<link rel="author" title="Chris Nardi" href="mailto:cnardi@chromium.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="container">
  <div id="target"></div>
</div>
<style>
  #container {
    color: rgb(255, 0, 0);
  }
</style>
<script>
tests = [
    ["/**/transparent", "rgba(0, 0, 0, 0)", "Should parse to completely transparent"],
    ["transparent", "rgba(0, 0, 0, 0)", "Should parse to completely transparent"],
    [" transparent\n", "rgba(0, 0, 0, 0)", "Should parse to completely transparent"],
    ["TransParent", "rgba(0, 0, 0, 0)", "Should parse to completely transparent"],
    ["currentColor", "rgb(255, 0, 0)", "Should be same as parent color"],
    ["CURRENTcolor", "rgb(255, 0, 0)", "Should be same as parent color"],
    ["black", "rgb(0, 0, 0)", "Should parse as correct value"],
    ["white", "rgb(255, 255, 255)", "Should parse as correct value"],
    ["fuchsia", "rgb(255, 0, 255)", "Should parse as correct value"],
    ["cyan", "rgb(0, 255, 255)", "Should parse as correct value"],
    ["CyAn", "rgb(0, 255, 255)", "Should parse as cyan"],
];

for (const test of tests) {
    test_computed_value("color", test[0], test[1], test[2] ? `[${test[2]}]` : undefined);
}
</script>
</body>
</html>
